Backend/requirements.txt
fastapi
uvicorn
opencv-python
moviepy
SpeechRecognition
transformers
torch
pillow
ollama
faiss-cpu
google-generativeai
wikipedia
numpy 